requirements.txt
aiogram==3.10.0
aiosqlite==0.19.0
python-dotenv==1.0.0
aiohttp==3.9.1